molecular_weight/requirements.txt
streamlit
chempy
rdkit-pypi